refactor(data): selectData 파일 분리
